move examples out of the library
