+++ .gitignore
@@ -1,4 +1,4 @@
-feedly/examples/results/*
+examples/results/*
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
