+++ examples/setup_auth.py
@@ -1,0 +1,4 @@
+from feedly.api_client.utils import setup_auth
+
+if __name__ == "__main__":
+    setup_auth(overwrite=True)
